+++ docs/history.rst
@@ -4,6 +4,10 @@ Version History
   * introduced: 24-bit color support, detected by ``term.number_of_colors == 1 << 24``,
     and 24-bit color foreground method :meth:`~Terminal.color_rgb` and background method
     :meth:`~Terminal.on_color_rgb`.
+  * bugfix: off-by-one error in :meth:`~.Terminal.get_location`, now accounts for
+    ``%i`` in cursor_report, :ghissue:`94`.
+  * bugfix :meth:`~Terminal.split_seqs` and related functions failed to match when the
+    color index was greater than 15, :ghissue:`101`.
   * bugfix: Context Managers, :meth:`~.Terminal.fullscreen`,
     :meth:`~.Terminal.hidden_cursor`, and :meth:`~Terminal.keypad`
     now flush the stream after writing their sequences.